requirements.txt
django
requests
djangorestframework_simplejwt
django-rest-framework

